Move root restore flow to backups dashboard
